requirements_test.txt
colorlog
PyYAML
elasticsearch
influxdb
requests
shellbot
urllib3[secure]
vcrpy
xmltodict
